requirements.txt
numpy
soundfile
scipy
faster-whisper
openai
python-dotenv
